% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{arad_parse_date}
\alias{arad_parse_date}
\title{Convert ARAD period to Date}
\usage{
arad_parse_date(period_string)
}
\arguments{
\item{period_string}{Character vector of period strings}
}
\value{
Date vector
}
\description{
Helper function to convert ARAD period format (YYYYMMDD) to Date object
}
\examples{
\dontrun{
periods <- c("20231201", "20231101")
dates <- arad_parse_date(periods)
}
}
